+++ mne/layouts/__init__.py
@@ -1,1 +1,1 @@
-from .layout import Layout
+from .layout import Layout, make_eeg_layout, read_layout