+++ mne/proj.py
@@ -5,8 +5,9 @@
 import numpy as np
 from scipy import linalg
 
-from . import fiff
+from . import fiff, Epochs
 from .fiff.pick import pick_types
+from .event import make_fixed_length_events
 
 
 def read_proj(fname):
@@ -133,3 +134,63 @@ def compute_proj_evoked(evoked, n_grad=2, n_mag=2, n_eeg=2):
     data = np.dot(evoked.data, evoked.data.T)  # compute data covariance
     desc_prefix = "%-.3f-%-.3f" % (evoked.times[0], evoked.times[-1])
     return _compute_proj(data, evoked.info, n_grad, n_mag, n_eeg, desc_prefix)
+
+
+def compute_proj_raw(raw, start=0, stop=None, duration=1, n_grad=2, n_mag=2,
+                     n_eeg=0, reject=None, flat=None):
+    """Compute SSP (spatial space projection) vectors on Raw
+
+    Parameters
+    ----------
+    raw: instance of Raw
+        A raw object to use the data from
+    start: float
+        Time (in sec) to start computing SSP
+    stop: float
+        Time (in sec) to stop computing SSP
+        None will go to the end of the file
+    duration: float
+        Duration (in sec) to chunk data into for SSP
+        If duration is None, data will not be chunked.
+    n_grad: int
+        Number of vectors for gradiometers
+    n_mag: int
+        Number of vectors for gradiometers
+    n_eeg: int
+        Number of vectors for gradiometers
+    reject : dict
+        Epoch rejection configuration (see Epochs)
+    flat : dict
+        Epoch flat configuration (see Epochs)
+
+    Returns
+    -------
+    projs: list
+        List of projection vectors
+    """
+    if duration is not None:
+        events = make_fixed_length_events(raw, 999, start, stop, duration)
+        epochs = Epochs(raw, events, None, tmin=0., tmax=duration,
+                        picks=pick_types(raw.info, meg=True, eeg=True),
+                        reject=reject, flat=flat)
+        data = sum(np.dot(e, e.T) for e in epochs)  # compute data covariance
+        if not stop:
+            stop = (raw.last_samp - raw.first_samp + 1) / raw.info['sfreq']
+    else:
+        # convert to sample indices
+        start = raw.time_to_index(start)
+        start = start[0]
+        if stop:
+            stop = raw.time_to_indx(stop)
+            stop = min(stop[0], raw.last_samp - raw.fist_samp + 1)
+        else:
+            stop = raw.last_samp - raw.first_samp + 1
+        data, times = raw[:, start:stop]
+        data = np.dot(data, data.T)  # compute data covariance
+        # convert back to times
+        start = start / raw.info['sfreq']
+        stop = stop / raw.info['sfreq']
+
+    desc_prefix = "Raw-%-.3f-%-.3f" % (start, stop)
+    projs = _compute_proj(data, raw.info, n_grad, n_mag, n_eeg, desc_prefix)
+    return projs